+++ bee-storage/bee-storage-memory/src/access/insert.rs
@@ -13,7 +13,11 @@ use bee_message::{
     output::OutputId,
     Message, MessageId,
 };
-use bee_storage::{access::Insert, backend::StorageBackend, system::System};
+use bee_storage::{
+    access::{Insert, InsertStrict},
+    backend::StorageBackend,
+    system::System,
+};
 use bee_tangle::{
     metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
@@ -34,7 +38,6 @@ macro_rules! impl_insert {
 
 impl_insert!(u8, System, system);
 impl_insert!(MessageId, Message, message_id_to_message);
-impl_insert!(MessageId, MessageMetadata, message_id_to_metadata);
 impl_insert!((MessageId, MessageId), (), message_id_to_message_id);
 impl_insert!(OutputId, CreatedOutput, output_id_to_created_output);
 impl_insert!(OutputId, ConsumedOutput, output_id_to_consumed_output);
@@ -52,3 +55,17 @@ impl_insert!(
 );
 impl_insert!((MilestoneIndex, Receipt), (), milestone_index_to_receipt);
 impl_insert!((bool, TreasuryOutput), (), spent_to_treasury_output);
+
+impl InsertStrict<MessageId, MessageMetadata> for Storage {
+    fn insert_strict(&self, k: &MessageId, v: &MessageMetadata) -> Result<(), <Self as StorageBackend>::Error> {
+        let mut guard = self.inner.write()?;
+
+        if !guard.message_id_to_metadata.exist(k) {
+            guard.message_id_to_metadata.insert(k, v);
+        }
+
+        drop(guard);
+
+        Ok(())
+    }
+}