+++ bee-storage/bee-storage-sled/src/access/insert.rs
@@ -13,7 +13,11 @@ use bee_message::{
     output::OutputId,
     Message, MessageId,
 };
-use bee_storage::{access::Insert, backend::StorageBackend, system::System};
+use bee_storage::{
+    access::{Insert, InsertStrict},
+    backend::StorageBackend,
+    system::System,
+};
 use bee_tangle::{
     metadata::MessageMetadata, solid_entry_point::SolidEntryPoint, unreferenced_message::UnreferencedMessage,
 };
@@ -39,15 +43,19 @@ impl Insert<MessageId, Message> for Storage {
     }
 }
 
-impl Insert<MessageId, MessageMetadata> for Storage {
-    fn insert(
+impl InsertStrict<MessageId, MessageMetadata> for Storage {
+    fn insert_strict(
         &self,
         message_id: &MessageId,
         metadata: &MessageMetadata,
     ) -> Result<(), <Self as StorageBackend>::Error> {
         self.inner
             .open_tree(TREE_MESSAGE_ID_TO_METADATA)?
-            .insert(message_id, metadata.pack_to_vec())?;
+            .update_and_fetch(message_id, |old_metadata| {
+                old_metadata
+                    .map(|b| b.to_vec())
+                    .or_else(|| Some(metadata.pack_to_vec()))
+            })?;
 
         Ok(())
     }